Moved setup 3D to headset module.
The plan is to move more and more of viewport into the headset module.
